Marketplace works 1.0

+++ src/main/java/com/mycompany/heromarsspring/services/MarketService.java
@@ -12,6 +12,8 @@ import com.mycompany.heromarsspring.entities.Hero;
 import com.mycompany.heromarsspring.entities.Item;
 import com.mycompany.heromarsspring.entities.ItemMarket;
 
+import lombok.Getter;
+
 @Service
 public class MarketService {
 
@@ -23,19 +25,26 @@ public class MarketService {
 
 	@Autowired
 	private HeroRepository heroRepository;
+	
+	@Getter
+	private String message;
 
 	public String saveItem(Integer itemId, Integer price) {
 
 		Item item = itemRepository.findByItemId(itemId);
 		ItemMarket market = new ItemMarket();
+		
+		Integer calculatedPrice = (int)(price * item.getLevel() * (item.getDurability()/100.0));
 
-		market.setPrice(price);
+		market.setPrice(calculatedPrice);
 		item.setMarketPresence(market);
 
 		marketRepo.saveAndFlush(market);
+		
+		message = "Sikeresen kitetted a " + item.getName().getDescription() + " eszközödet a piacra " + calculatedPrice
+				+ " buznyák áron!"; 
 
-		return "Sikeresen kitetted a " + item.getName().getDescription() + "eszközödet a piacra " + price
-				+ "buznyák áron";
+		return message;
 	}
 
 	public List<ItemMarket> getItems() {
@@ -53,22 +62,32 @@ public class MarketService {
 	public List<ItemMarket> buyItem(Integer id, String heroName) {
 
 		ItemMarket itemMarket = marketRepo.getOne(id);
-
 		Item item = itemMarket.getItem();
-
 		Hero sellerHero = item.getHero();
-
 		Hero buyerHero = heroRepository.findByHeroName(heroName);
 
-		buyerHero.setMoney(buyerHero.getMoney() - itemMarket.getPrice());
-		item.setHero(buyerHero);
-
-		sellerHero.setMoney(sellerHero.getMoney() + itemMarket.getPrice());
-
-		item.setMarketPresence(null);
-
-		marketRepo.deleteById(id);
-
-		return marketRepo.findAll();
+		if (buyerHero.getMoney()<itemMarket.getPrice()) {
+			
+			message = "Nincs elegendő buznyákod megvásárolni ezt a " 
+					+ item.getName().getDescription() + "  eszközt!";
+			
+			return marketRepo.findAll();
+			
+		} else {
+		
+			buyerHero.setMoney(buyerHero.getMoney() - itemMarket.getPrice());
+			item.setHero(buyerHero);
+	
+			sellerHero.setMoney(sellerHero.getMoney() + itemMarket.getPrice());
+	
+			item.setMarketPresence(null);
+	
+			marketRepo.deleteById(id);
+			
+			message ="Sikeres megvásároltad a " + item.getName().getDescription() + " eszközt a " 
+					+ sellerHero.getHeroName() + " nevű hőstől!";
+	
+			return marketRepo.findAll();
+		}
 	}
 }
